+++ lib/core/api/better_auth_client.g.dart
@@ -604,6 +604,42 @@ class _BetterAuthClient implements BetterAuthClient {
     );
   }
 
+  Future<HttpResponse<StatusResponse>> _verifyPassword({
+    required String password,
+  }) async {
+    final _extra = <String, dynamic>{};
+    final queryParameters = <String, dynamic>{};
+    final _headers = <String, dynamic>{};
+    final _data = <String, dynamic>{'password': password};
+    final _options = _setStreamType<Result<StatusResponse>>(
+      Options(method: 'POST', headers: _headers, extra: _extra)
+          .compose(
+            _dio.options,
+            '/verify-password',
+            queryParameters: queryParameters,
+            data: _data,
+          )
+          .copyWith(baseUrl: _combineBaseUrls(_dio.options.baseUrl, baseUrl)),
+    );
+    final _result = await _dio.fetch<Map<String, dynamic>>(_options);
+    late StatusResponse _value;
+    try {
+      _value = StatusResponse.fromJson(_result.data!);
+    } on Object catch (e, s) {
+      errorLogger?.logError(e, s, _options, response: _result);
+      rethrow;
+    }
+    final httpResponse = HttpResponse(_value, _result);
+    return httpResponse;
+  }
+
+  @override
+  Future<Result<StatusResponse>> verifyPassword({required String password}) {
+    return BetterAuthCallAdapter<StatusResponse>().adapt(
+      () => _verifyPassword(password: password),
+    );
+  }
+
   RequestOptions _setStreamType<T>(RequestOptions requestOptions) {
     if (T != dynamic &&
         !(requestOptions.responseType == ResponseType.bytes ||