+++ lib/flutter_better_auth.dart
@@ -1,6 +1,7 @@
 export 'core/api/default/sign_in/models/email/sign_in_email_response.dart';
 export 'core/api/default/sign_in/models/social/sign_in_social_response.dart';
 export 'core/api/default/sign_in/models/social/social_id_token_body.dart';
+export 'core/api/default/sign_in/models/username/username_available_response.dart';
 export 'core/api/default/sign_in/sign_in_better_auth.dart';
 
 // DEFAULT PLUGINS